% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Bi_Draw_ConfuMax.R
\name{Bi_CMPlotSim}
\alias{Bi_CMPlotSim}
\title{A function for drawing confusion matrix in simple version}
\usage{
Bi_CMPlotSim(cm, TestID = "Predict1")
}
\arguments{
\item{cm}{Dataframe of confusion matrix.}

\item{TestID}{Title name of the plot.}
}
\description{
This function allows you to draw confusion matrix for binary data in simple version.
}
\examples{
Draw_Bi_CM(cm,TestID = "Predict1")
}
\keyword{confusion}
\keyword{matrix}
