% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_Bi_SummarizeCM.R
\name{Bi_SummarizeCM}
\alias{Bi_SummarizeCM}
\title{A function for summarizing the results from multiple binary confusion matrix to dataframe}
\usage{
Bi_SummarizeCM(cm.lt, Anno.df)
}
\arguments{
\item{cm.lt}{A list of multiple binary confusion matrix.}

\item{Anno.df}{A dataframe of the annotation table.}
}
\description{
This function allows you to summarize the results from multiple binary confusion matrix to dataframe.
}
\examples{
Bi_SummarizeCM(cm.lt, Anno.df)

}
\keyword{confusion}
\keyword{matrix}
\keyword{summarize}
\keyword{the}
