% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_CellCheck_Bi.R
\name{CellCheck_Bi}
\alias{CellCheck_Bi}
\title{A function for creating summarize results of binary data.}
\usage{
CellCheck_Bi(
  Bi.df = Simu_Bi.df,
  Anno.df = Simu_Anno.df,
  CMSet.lt = "",
  BarChartSet.lt = "",
  LinePlotSet.lt = "",
  Save.Path = "",
  ProjectName = ""
)
}
\arguments{
\item{Bi.df}{A binary dataframe of answers and results.}

\item{Anno.df}{A dataframe of annotation.}

\item{CMSet.lt}{Confusion matrix setting.
Mode = c("One","Multiple"), Mode = "One": Prediction selection.; Mode = "Multiple": All Predictions.;
Actual: Correct answer setting.; Predict: Prediction selection in Mode = "One".; Remark = PDF file name setting.}

\item{BarChartSet.lt}{Bar Chart setting.
#' Mode = c("One","Multiple"), Mode = "One": Metric selection for y-axis setting.; Mode = "Multiple": All Metrics.;
Metric: Metric selection in Mode = "One" for y-axis setting.; XValue: x-axis setting.; Group: Group setting.; Remark = PDF file name setting.}

\item{LinePlotSet.lt}{Line Plot setting.
#' Mode = c("One","Multiple"), Mode = "One": Metric selection for y-axis setting.; Mode = "Multiple": All Metrics.;
Metric: Metric selection in Mode = "One" for y-axis setting.; XValue: x-axis setting.; Group: Group setting.; Remark = PDF file name setting.}

\item{Save.Path}{The setting of the saving path.Defaults to the path of the scripts folder.}

\item{ProjectName}{The naming of project Name.}
}
\description{
This function allows you to create summarize results from binary data.
}
\examples{
CellCheck_Bi(Simu_Bi.df, Simu_Anno.df,
             CMSet.lt = CMSet.lt,
             BarChartSet.lt = BarChartSet.lt,
             LinePlotSet.lt = LinePlotSet.lt,
             Save.Path="", ProjectName="")

}
\keyword{Summarize}
\keyword{binary}
\keyword{data.}
\keyword{of}
\keyword{results}
