% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_CellCheck_Conti.R
\name{CellCheck_Conti}
\alias{CellCheck_Conti}
\title{A function for creating summarize results of continuous data.}
\usage{
CellCheck_Conti(
  Conti.df = Sum_Conti.df,
  Anno.df = Simu_Anno.df,
  Mode = "Multiple",
  BarMetricSet.lt,
  Save.Path = "",
  ProjectName = ""
)
}
\arguments{
\item{Anno.df}{A dataframe of annotation.}

\item{Mode}{one of 'One' or 'Multiple'. Chose Mode = "One" will export the result of the chosen index by the setting of BarMetricSet.lt, and chose Mode = "Multiple" will export all indexes in the dataframe and integrated results.}

\item{BarMetricSet.lt}{Set the index for subsequent comparison in Mode = "one".}

\item{Save.Path}{The setting of the saving path.Defaults to the path of the scripts folder.}

\item{ProjectName}{The naming of project Name.}

\item{Conti.df.df}{A dataframe of continuous answers and results.}
}
\description{
This function allows you to create summarize results from continuous data.
}
\examples{
CellCheck_Conti(Conti.df = Sum_Conti.df, Anno.df = Simu_Anno.df,
                Mode = "Multiple", BarMetricSet.lt,
                Save.Path="", ProjectName="")

}
\keyword{Summarize}
\keyword{continuous}
\keyword{data.}
\keyword{of}
\keyword{results}
