% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_CellCheck_DisMult.R
\name{CellCheck_DisMult}
\alias{CellCheck_DisMult}
\title{A function for creating summarize results of multiple discrete data.}
\usage{
CellCheck_DisMult(
  Simu_DisMult.df,
  Simu_Anno.df,
  DisCMSet.lt = "",
  BarChartSet.lt = "",
  LinePlotSet.lt = "",
  Save.Path = "",
  ProjectName = ""
)
}
\arguments{
\item{DisCMSet.lt}{Confusion matrix setting.
Mode = c("One","Multiple"), Mode = "One": Prediction selection; Mode = "Multiple": All Predictions;
FilterSet1 = Select the category; FilterSet2 = Select the type in FilterSet1;
Actual: Correct answer setting; Predict: Prediction selection in Mode = "One"; Remark = PDF file name setting.}

\item{BarChartSet.lt}{Bar Chart setting.
Mode = c("One","Multiple"), Mode = "One": Metric selection for y-axis setting; Mode = "Multiple": All Metrics;
Metric: Metric selection in Mode = "One" for y-axis setting; XValue: x-axis setting; Group: Group setting; Remark = PDF file name setting.}

\item{LinePlotSet.lt}{Line Plot setting.
Mode = c("One","Multiple"), Mode = "One": Metric selection for y-axis setting; Mode = "Multiple": All Metrics;
Metric: Metric selection in Mode = "One" for y-axis setting; XValue: x-axis setting; Group: Group setting; Remark = PDF file name setting.}

\item{Save.Path}{The setting of the saving path.Defaults to the path of the scripts folder.}

\item{ProjectName}{The naming of project Name.}

\item{DisMult.df}{A multiple discrete dataframe of answers and results.}

\item{Anno.df}{A dataframe of annotation.}
}
\description{
This function allows you to create summarize results from multiple discrete data.
}
\examples{
CellCheck_DisMult(Simu_DisMult.df, Simu_Anno.df,
                  DisCMSet.lt  = DisCMSet.lt ,
                  BarChartSet.lt = BarChartSet.lt,
                  LinePlotSet.lt = LinePlotSet.lt,
                  Save.Path="", ProjectName="")

}
\keyword{Summarize}
\keyword{data.}
\keyword{discrete}
\keyword{multiple}
\keyword{of}
\keyword{results}
