% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_DiscMult_ACC_ER.R
\name{DiscMult_AccEr}
\alias{DiscMult_AccEr}
\title{A function for creating Measure accuracy dataframe}
\usage{
DiscMult_AccEr(Simu_DisMult.df, Simu_Anno.df)
}
\arguments{
\item{Simu_Anno.df}{A dataframe of the annotation table.}

\item{Simu_DisMult.d}{A dataframe of the discrete multiple data.}
}
\description{
This function allows you to create dataframe of measure accuracy from discrete multiple data.
}
\examples{
Measure_Accuracy(Actual, Predict,  NaRm = FALSE)

}
\keyword{confusion}
\keyword{matrix}
\keyword{summarize}
\keyword{the}
