% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_Bi_Draw_ConfuMatrix.R
\name{Draw_Bi_CM}
\alias{Draw_Bi_CM}
\title{A function for drawing confusion matrix}
\usage{
Draw_Bi_CM(cm, TestID = "Predict1")
}
\arguments{
\item{cm}{Dataframe of confusion matrix.}

\item{TestID}{Title name of the plot.}
}
\description{
This function allows you to draw confusion matrix for binary data.
}
\examples{
Draw_Bi_CM(cm,TestID = "Predict1")
}
\keyword{confusion}
\keyword{matrix}
