% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_Conti_Measure_Accuracy.R
\name{Measure_Accuracy}
\alias{Measure_Accuracy}
\title{A function for creating Measure accuracy dataframe}
\usage{
Measure_Accuracy(Actual, Predict, NaRm = FALSE)
}
\arguments{
\item{Actual}{A serous of actual value.}

\item{Predict}{A serous of predict value.}
}
\description{
This function allows you to create dataframe of measure accuracy from continuous data .
}
\examples{
Measure_Accuracy(Actual, Predict,  NaRm = FALSE)

}
\keyword{confusion}
\keyword{matrix}
\keyword{summarize}
\keyword{the}
