% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_CellCheck_Bi.R
\name{CellCheck_Bi}
\alias{CellCheck_Bi}
\title{A function for creating summarize results of binary data.}
\usage{
CellCheck_Bi(
  Bi.df = Simu_Bi.df,
  Anno.df = Simu_Anno.df,
  Mode = "Multiple",
  CMPredSet.lt,
  Save.Path = "",
  ProjectName = ""
)
}
\arguments{
\item{Bi.df}{A binary dataframe of answers and results.}

\item{Anno.df}{A dataframe of annotation.}

\item{Mode}{one of 'One' or 'Multiple'. Chose Mode = "One" will export the confusion matrix of chosen prediction result by the setting of CMPredSet.lt, and chose Mode = "Multiple" will export all predictions in the dataframe and integrated results.}

\item{CMPredSet.lt}{Set correct answers and predicted results for subsequent comparison in Mode = "one".}

\item{Save.Path}{The setting of the saving path.Defaults to the path of the scripts folder.}

\item{ProjectName}{The naming of project Name.}
}
\description{
This function allows you to create summarize results from binary data.
}
\examples{
CellCheck_Bi(Simu_Bi.df, Simu_Anno.df,
             Mode = "Multiple", CMPredSet.lt,
             Save.Path="", ProjectName="")

}
\keyword{Summarize}
\keyword{binary}
\keyword{data.}
\keyword{of}
\keyword{results}
