% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_CellCheck_Conti.R
\name{CellCheck_Conti}
\alias{CellCheck_Conti}
\title{A function for creating summarize results of continuous data.}
\usage{
CellCheck_Conti(
  Conti.df = Sum_Conti.df,
  Anno.df = Simu_Anno.df,
  BarChartSet.lt = "",
  LinePlotSet.lt = "",
  Save.Path = "",
  ProjectName = ""
)
}
\arguments{
\item{Anno.df}{A dataframe of annotation.}

\item{BarChartSet.lt}{Bar Chart setting.
Mode = c("One","Multiple"), Mode = "One": Metric selection for y-axis setting; Mode = "Multiple": All Metrics;
Metric: Metric selection in Mode = "One" for y-axis setting; XValue: x-axis setting; Group: Group setting; Remark = PDF file name setting.}

\item{LinePlotSet.lt}{Line Plot setting.
Mode = c("One","Multiple"), Mode = "One": Metric selection for y-axis setting; Mode = "Multiple": All Metrics;
Metric: Metric selection in Mode = "One" for y-axis setting; XValue: x-axis setting; Group: Group setting; Remark = PDF file name setting.}

\item{Save.Path}{The setting of the saving path.Defaults to the path of the scripts folder.}

\item{ProjectName}{The naming of project Name.}

\item{Conti.df.df}{A dataframe of continuous answers and results.}
}
\description{
This function allows you to create summarize results from continuous data.
}
\examples{
CellCheck_Conti(Conti.df = Sum_Conti.df, Anno.df = Simu_Anno.df,
                BarChartSet.lt = BarChartSet.lt,
                LinePlotSet.lt =  LinePlotSet.lt,
                Save.Path="", ProjectName="")

}
\keyword{Summarize}
\keyword{continuous}
\keyword{data.}
\keyword{of}
\keyword{results}
