% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_CellCheck_DisMult.R
\name{CellCheck_DisMult}
\alias{CellCheck_DisMult}
\title{A function for creating summarize results of multiple discrete data.}
\usage{
CellCheck_DisMult(
  Simu_DisMult.df,
  Simu_Anno.df,
  Mode = "Multiple",
  DisMultCM.lt,
  Save.Path = "",
  ProjectName = ""
)
}
\arguments{
\item{Mode}{one of 'One' or 'Multiple'. Chose Mode = "One" will export the confusion matrix of chosen prediction result by the setting of DisMultCM.lt, and chose Mode = "Multiple" will export all predictions in the dataframe and integrated results.}

\item{DisMultCM.lt}{Set correct answers and predicted results for subsequent comparison in Mode = "one".}

\item{Save.Path}{The setting of the saving path.Defaults to the path of the scripts folder.}

\item{ProjectName}{The naming of project Name.}

\item{DisMult.df}{A multiple discrete dataframe of answers and results.}

\item{Anno.df}{A dataframe of annotation.}
}
\description{
This function allows you to create summarize results from multiple discrete data.
}
\examples{
CellCheck_Bi(Simu_DisMult.df, Simu_Anno.df,
             Mode = "Multiple", DisMultCM.lt,
             Save.Path="", ProjectName="")

}
\keyword{Summarize}
\keyword{data.}
\keyword{discrete}
\keyword{multiple}
\keyword{of}
\keyword{results}
